Move all dynamic file support to VS layer
