+++ src/Razor/src/Microsoft.VisualStudio.Editor.Razor/DynamicFiles/BackgroundDocumentGenerator.cs
@@ -3,42 +3,34 @@
 
 using System;
 using System.Collections.Generic;
-using System.Composition;
+using System.ComponentModel.Composition;
 using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Razor;
-using Microsoft.CodeAnalysis.Razor.DynamicFiles;
+using Microsoft.CodeAnalysis.Razor;
 using Microsoft.CodeAnalysis.Razor.ProjectSystem;
 using Microsoft.Extensions.Internal;
 using Microsoft.VisualStudio.Threading;
 
-namespace Microsoft.CodeAnalysis.Razor;
+namespace Microsoft.VisualStudio.Razor.DynamicFiles;
 
-[Shared]
 [Export(typeof(IProjectSnapshotChangeTrigger))]
-internal class BackgroundDocumentGenerator : IProjectSnapshotChangeTrigger
+[method: ImportingConstructor]
+internal class BackgroundDocumentGenerator(
+    ProjectSnapshotManagerDispatcher dispatcher,
+    IRazorDynamicFileInfoProviderInternal infoProvider) : IProjectSnapshotChangeTrigger
 {
     // Internal for testing
-    internal readonly Dictionary<DocumentKey, (IProjectSnapshot project, IDocumentSnapshot document)> Work;
+    internal readonly Dictionary<DocumentKey, (IProjectSnapshot project, IDocumentSnapshot document)> Work = [];
 
-    private readonly ProjectSnapshotManagerDispatcher _dispatcher;
-    private readonly IRazorDynamicFileInfoProviderInternal _infoProvider;
-    private readonly HashSet<string> _suppressedDocuments;
+    private readonly ProjectSnapshotManagerDispatcher _dispatcher = dispatcher;
+    private readonly IRazorDynamicFileInfoProviderInternal _infoProvider = infoProvider;
+    private readonly HashSet<string> _suppressedDocuments = new(FilePathComparer.Instance);
     private ProjectSnapshotManagerBase? _projectManager;
     private Timer? _timer;
     private bool _solutionIsClosing;
-
-    [ImportingConstructor]
-    public BackgroundDocumentGenerator(ProjectSnapshotManagerDispatcher dispatcher, IRazorDynamicFileInfoProviderInternal infoProvider)
-    {
-        _dispatcher = dispatcher ?? throw new ArgumentNullException(nameof(dispatcher));
-        _infoProvider = infoProvider ?? throw new ArgumentNullException(nameof(infoProvider));
-        _suppressedDocuments = new HashSet<string>(FilePathComparer.Instance);
-
-        Work = new Dictionary<DocumentKey, (IProjectSnapshot project, IDocumentSnapshot document)>();
-    }
 
     public bool HasPendingNotifications
     {
