+++ src/Razor/src/Microsoft.VisualStudio.Editor.Razor/DynamicFiles/ExcerptResultInternal.cs
@@ -2,11 +2,12 @@
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
 using System.Collections.Immutable;
+using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Classification;
 using Microsoft.CodeAnalysis.ExternalAccess.Razor;
 using Microsoft.CodeAnalysis.Text;
 
-namespace Microsoft.CodeAnalysis.Razor.DynamicFiles;
+namespace Microsoft.VisualStudio.Razor.DynamicFiles;
 
 // We have IVT access to the Roslyn APIs for product code, but not for testing.
 internal readonly struct ExcerptResultInternal
