+++ src/Razor/src/Microsoft.VisualStudio.Editor.Razor/DynamicFiles/RazorDocumentServiceProviderFactory.cs
@@ -2,12 +2,11 @@
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
 using System;
-using System.Composition;
+using System.ComponentModel.Composition;
 using Microsoft.CodeAnalysis.ExternalAccess.Razor;
 
-namespace Microsoft.CodeAnalysis.Razor.DynamicFiles;
+namespace Microsoft.VisualStudio.Razor.DynamicFiles;
 
-[Shared]
 [Export(typeof(IRazorDocumentServiceProviderFactory))]
 internal sealed class RazorDocumentServiceProviderFactory : IRazorDocumentServiceProviderFactory
 {
